Use consistent STATICFILES_STORAGE across environments.
collectstatic only adds hashes to static file filenames when
STATICFILES_STORAGE is set correctly. We want the Dockerfile and
other things to have the hashed filenames, but the Dockerfile runs
collectstatic under the Development environment. Also, there's no
problem using the manifest storage during development.

So we move the setting.