[fix bug 1262617] Replace RevisionMiddleware with decorators.
RevisionMiddleware automatically creates revisions when things
change. However, viewing the revert page in the admin interface
triggers a revert (in order to generate the preview data before
actually reverting) within a transaction, and then rolls back the
transaction. But, as far as I can tell, django-reversion doesn't
handle transaction rollback and saves the revision from the revert
anyway.

Removing the middleware fixes this, but we still have to manually
trigger revision creation outside the admin. This commit adds
decorators for this to the Action ViewSet.