Do not include hashes in client assets in development.
Currently, built assets have hashes in their filenames by default.
When working locally and using the watch command, this has the side
effect of generating a bunch of nearly-identical assets with different
filenames, and over time can take up a lot of space.

In a development environment there's no reason to use unique hashes in
asset filenames, so disabling this behavior will help avoid the buildup
of assets.