Do not use an XRay-ed Promise during recipe execution.
The RecipeRunner uses a promise from the sandbox to detect when
execution is done, but the promise methods aren't available since the
object is XRay-ed and its prototype is set to Object.

Instead, this patch uses two new functions passed to the sandbox that
map to resolving or rejecting a promise. The new executeAction function
returns a Promise defined outside the sandbox, but resolved or rejected
from within it when execution is finished.

The Promise also makes testing a bit easier as we can now inspect the
value that the sandbox resolves with.