Move EventEmitter out of the sandbox.
EventEmitter involves a bunch of hard-to-follow cloning due to it
running in the sandbox and taking the driver as an argument, but all it
uses the driver for is logging. To simplify this (and to solve a bug
around the cloning), this patch moves it outside of the sandbox and
removes the dependency on the driver.

The main ugly part is that, because it emits data cloned into the
sandbox, it can't freeze wrapped data it's given. To freeze it, it would
need a reference to the sandbox itself, which causes other issues when
the emitter is cloned into the sandbox. The non-ideal workaround for
now is to freeze the objects before cloning them.