Fix #549: Remove eventlet from requirements.
Requests uses eventlet for making async requests if it is available,
but the old version of eventlet we're using for Python 3.6 compatability
is, for some reason, very slow when used in the mock-recipe-server
tests. We only kept eventlet around for using eventlet workers in
gunicorn on prod, but we're considering other worker types instead.