Bug 1349596: Validate JEXL on the server instead of the client.
Validating JEXL on the client isn't ideal, because the JEXL library doesn't
have a public API for checking if a statement is valid without evaluating
it, which means certain valid expressions won't pass validation if they
reference data in the context that we don't provide in the admin interface.

PyJEXL is a Python library that parses and evaluates JEXL, and it includes
a validation API that is more suitable for our validation.