Convert PreferenceExperiments.jsm tests to async, and add stop test.
Because withMockPreferences expected a generator, but withMockExperiments
expected an async test functions, the tests for PreferenceExperiments.jsm
weren't actually being run properly, masking failures.

Converting the entire test suite to async fixes this issue, and revealed a
test failure in PreferenceExperiments.stop, which is also fixed here (and
we add a new test for the behavior that was failing).