recipe-server/contract-tests: Skip cache tests on api URLs.
To work around Firefox caching bugs, we are hiding the the cache
headers on API views in Nginx. These cache headers are still used
server side to cache requests on webheads, but sending them to Firefox
currently (on Beta 53) causes it to serve very stale requests.

This is temporary, and should be removed when the Firefox bug is fixed.