Fix #675: Split fetching and linting out of build task, and fix path.
Downloading gecko-dev is now a standalone task rather than part of the
build path. In addition, because linting does not require a built copy
of gecko-dev, it has been moved to a standalone task that depends on
the fetch task.

This also fixes the path given to update-mozilla-central.sh to point
to the correct directory.