Fetch tags from mozilla/normandy before syncing in TC fetch script.
TaskCluster uses the fork from a PR as the Git repo as the remote for
the checkout in TC jobs, which means that tags from the mozilla source
repo are not normally available. By adding and fetching the mozilla
repo to the checkout, we fetch the tags, which allows the sync script
to create a better version number for the add-on while it applies its
changes to the gecko-dev checkout.