#698: Treat malformed lastShown dates as valid in show-heartbeat.
Previously, we tried to parse the lastShown date when checking if the
user has ever seen a heartbeat prompt. If the parsing failed, we
treated the resulting value (NaN) as false, which triggered
reprompting the user.

With this change, we treat _any_ stored value for lastShown as meaning
that the user saw the prompt at least once. This avoids reprompting
users with buggy storage, as well as issues with our own code writing
incorrect values to localStorage.