Fix #718: Move add-on startup and shutdown to a new module.
To write unit tests for startup and shutdown logic, we need to import
the functions into a test file. We also cannot import files using
resource:// URLs that aren't in the jar after a build.

However, the build process for system add-ons includes bootstrap.js
in the jar already. So when bootstrap.js was added to jar.mn in order
to make it importable during tests, it caused it to be included twice
in the jar. This is detected as an error during the packaging step of
building Firefox.

Moving startup and shutdown to a standalone module allows us to
import and test freely, without breaking packaging.