Fix bug 1368703: Validate preference experiment slugs.
If we wanted to auto-trim slugs, we'd want to do it server-side to
avoid issues with other clients submitting non-compliant slugs. We don't
yet have a good solution for server-side transformations on a per-action
basis. To avoid addressing that large issue, for now we can just validate
the slugs going forward. Long-term we'll want to address per-action
customization in a better way.