Fix mock-server to work with signed actions.
Since actions are signed now, we can't modify the URLs in the actions since
that would invalidate the signature. Instead, we switch to using the Django
test client for making API responses, which allows us to temporarily modify the
CDN_URL setting to the URL prefix we need for a testcase.