Convert serializers to not use current_version properties
We keep the same wishy-washy stance in the v1 RecipeSerializer,
serializing the approved_revision if any and falling back to the
current_revision. In v1 we only really need to serialize enabled
recipes, so in principle we can force approved_revision for all these
properties, but this seemed like a smaller change.

We take a stronger stance in MinimalRecipeSerializer, enforcing that
recipes must be approved before they can be serialized in this
way. Because this is the serialization mechanism used when signing a
recipe, it is now impossible to sign recipes that have not been approved.

- Add an approver to all RecipeFactory(signed=True) calls to ensure
  they are approved so they can be signed.

- test_signature_is_cleared_if_autograph_unavailable used to rely on
  the fact that there was no approved_revision. Now there is, because of
  the above change, so tweak the test so it continues to pass.

- Convert uses_only_baseline_capabilities fields to use fancier DRF
  "source" (it can automatically call no-argument methods).

- Restore test for identicon_seed that I took out in #2018 now that I
  understand a little better how this works and how to make it do what I
  want.