Add JEXL filter object
This could be used by NDT to combine with other filter objects, so that we
can nest a JEXL filter inside a sticky or OR expression, without having to
bail out entirely. It also adds the ability to comment the reasons behind a
JEXL expression, and note any additional capabilities that may be needed.