Make recipe enable and disable operations non-atomic
This was originally added so that if there was a Remote Settings API error,
the database changes would be rolled back. However, this gives too much power
to the sync process. Instead we should record what the state should be in the
database, and if it doesn't sync immediately, then the background job that
periodically re-sync RS and the database will catch it.

This avoid confusion such as the API call to Remote Settings seeming to fail,
but Remote Settings actually committed the change. Having the potential for
the Remote Settings collection to be "ahead" of the database is more damaging
then having to wait for eventual consistency.